Made the number of urls to curl for the http status dynamic
Changed the script so the number of IPs to curl per loop can be dynamic
and specified as a url param e.g. ips=x where x is the number of ips to
generate and then curl the http status codes of. You can now do block
of 20 IPs for example or more for increased speed. The default is 5.